✨ Feat :  qr 스캔 로직 훅 생성 및 적용
